more packets using fixed size integers

+++ src/db/Peer.cpp
@@ -127,7 +127,7 @@ void CPeer::SetP2PPort(WORD wPort)
 	m_wP2PPort = wPort;
 }
 
-void CPeer::SetMaps(long * pl)
+void CPeer::SetMaps(int32_t * pl)
 {
 	thecore_memcpy(m_alMaps, pl, sizeof(m_alMaps));
 }
