+++ src/db/Peer.h
@@ -52,8 +52,8 @@ class CPeer : public CPeerBase
 	void	SetP2PPort(WORD wPort);
 	WORD	GetP2PPort() { return m_wP2PPort; }
 
-	void	SetMaps(long* pl);
-	long *	GetMaps() { return &m_alMaps[0]; }
+	void	SetMaps(int32_t* pl);
+	int32_t*	GetMaps() { return &m_alMaps[0]; }
 
 	bool	SetItemIDRange(TItemIDRangeTable itemRange);
 	bool	SetSpareItemIDRange(TItemIDRangeTable itemRange);
@@ -68,7 +68,7 @@ class CPeer : public CPeerBase
 	DWORD	m_dwUserCount;
 	WORD	m_wListenPort;	// 게임서버가 클라이언트를 위해 listen 하는 포트
 	WORD	m_wP2PPort;	// 게임서버가 게임서버 P2P 접속을 위해 listen 하는 포트
-	long	m_alMaps[32];	// 어떤 맵을 관장하고 있는가?
+	int32_t	m_alMaps[32];	// 어떤 맵을 관장하고 있는가?
 
 	TItemIDRangeTable m_itemRange;
 	TItemIDRangeTable m_itemSpareRange;
